requirements.txt
discord.py==1.7.3
Telethon==1.24.0
